refactor(ios): extract Apple Wallet views and QR card from DashboardView
Behavior-preserving move (DashboardView.swift 388 -> 257 lines):

- AppleWalletPassSheet, AppleWalletAddPassButton, and
  AppleWalletAddPassView move to Features/Dashboard/AppleWalletViews.swift
  (private -> internal since they now cross a file boundary; all other
  code verbatim).
- QRCodeCardView moves verbatim to Features/Dashboard/QRCodeCardView.swift
  (it is shared by DashboardView and VenueModeView).
- pbxproj wired with explicit entries.
