+++ apps/ios/Jovie/Features/Dashboard/DashboardView.swift
@@ -255,134 +255,3 @@ struct DashboardView: View {
     }
   }
 }
-
-private struct AppleWalletPassSheet: Identifiable {
-  let id = UUID()
-  let pass: PKPass
-}
-
-private struct AppleWalletAddPassButton: UIViewRepresentable {
-  let isEnabled: Bool
-  let action: @MainActor () -> Void
-
-  func makeCoordinator() -> Coordinator {
-    Coordinator(action: action)
-  }
-
-  func makeUIView(context: Context) -> PKAddPassButton {
-    let button = PKAddPassButton(addPassButtonStyle: .black)
-    button.addTarget(
-      context.coordinator,
-      action: #selector(Coordinator.didTap),
-      for: .touchUpInside
-    )
-    return button
-  }
-
-  func updateUIView(_ button: PKAddPassButton, context: Context) {
-    context.coordinator.action = action
-    button.isUserInteractionEnabled = isEnabled
-    button.alpha = isEnabled ? 1 : 0.6
-  }
-
-  final class Coordinator: NSObject {
-    var action: @MainActor () -> Void
-
-    init(action: @escaping @MainActor () -> Void) {
-      self.action = action
-    }
-
-    @MainActor
-    @objc func didTap() {
-      action()
-    }
-  }
-}
-
-private struct AppleWalletAddPassView: UIViewControllerRepresentable {
-  let pass: PKPass
-
-  func makeCoordinator() -> Coordinator {
-    Coordinator()
-  }
-
-  func makeUIViewController(context: Context) -> UIViewController {
-    guard let controller = PKAddPassesViewController(pass: pass) else {
-      return UIViewController()
-    }
-    controller.delegate = context.coordinator
-    return controller
-  }
-
-  func updateUIViewController(_ controller: UIViewController, context: Context) {}
-
-  final class Coordinator: NSObject, PKAddPassesViewControllerDelegate {
-    func addPassesViewControllerDidFinish(_ controller: PKAddPassesViewController) {
-      controller.dismiss(animated: true)
-    }
-  }
-}
-
-/// A fixed-aspect QR card that reserves its square footprint up front and
-/// renders the QR off the main thread. The reserved space guarantees zero
-/// layout shift between the loading, loaded, and unavailable states. Shared by
-/// the dashboard and Venue Mode so both render identically. Uses the shared
-/// ``JovieQRCodePlate`` tokens so the plate matches the rest of the system.
-struct QRCodeCardView: View {
-  let payload: String?
-  var accessibilityLabelText: String = "Profile QR Code"
-
-  @State private var image: UIImage?
-
-  init(payload: String?, accessibilityLabelText: String = "Profile QR Code") {
-    self.payload = payload
-    self.accessibilityLabelText = accessibilityLabelText
-    // Seed from the renderer cache so a QR rendered once (e.g. on the
-    // dashboard) appears instantly when Venue Mode opens — no flash.
-    _image = State(
-      initialValue: payload.flatMap { QRCodeRenderer.cachedImage(for: $0) }
-    )
-  }
-
-  var body: some View {
-    ZStack {
-      RoundedRectangle(cornerRadius: JovieQRCodePlate.radius, style: .continuous)
-        .fill(payload == nil ? JovieColor.surface1 : JovieQRCodePlate.background)
-
-      if let image {
-        Image(uiImage: image)
-          .interpolation(.none)
-          .resizable()
-          .scaledToFit()
-          .padding(JovieQRCodePlate.padding)
-          .transition(.opacity)
-          .accessibilityLabel(accessibilityLabelText)
-      } else if payload == nil {
-        Text("QR unavailable")
-          .font(JovieFont.body(size: 15, weight: .medium))
-          .foregroundStyle(JovieColor.textTertiary)
-      }
-    }
-    .aspectRatio(1, contentMode: .fit)
-    .frame(maxWidth: .infinity)
-    .animation(JovieMotion.easeOut(duration: JovieMotion.subtleDuration), value: image == nil)
-    .task(id: payload) { await render() }
-  }
-
-  @MainActor
-  private func render() async {
-    guard let payload, !payload.isEmpty else {
-      image = nil
-      return
-    }
-
-    if let cached = QRCodeRenderer.cachedImage(for: payload) {
-      if image == nil {
-        image = cached
-      }
-      return
-    }
-
-    image = await QRCodeRenderer.imageAsync(for: payload)
-  }
-}
